third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/identity.https.any.sharedworker_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] identity float32 0D scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test identity float32 actual -4.2734375 should be close enough to expected -4.273642539978027 by ULP distance: expected a number less than or equal to 0n but got 430n
[FAIL] [required] identity float32 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by ULP distance: expected a number less than or equal to 0n but got 3691n
[FAIL] [required] identity float32 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by ULP distance: expected a number less than or equal to 0n but got 3691n
[FAIL] [required] identity float32 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by ULP distance: expected a number less than or equal to 0n but got 3691n
[FAIL] [required] identity float32 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by ULP distance: expected a number less than or equal to 0n but got 3691n
Harness: the test ran to completion.
